Remove travelling actors from collision hash before importing properties
This fixes a crash that occurs when a travelling actor has bCollideActors true in defaultproperties, but it was false at the time of travel. When the actor travels, the newly spawned actor in the destination level will have bCollideActors set to true from defaultproperties (and would therefore be present in the collision hash). Once the properties from the old actor are copied to it, bCollideActors will become false without removing the actor from the collision hash. This causes the newly spawned actor to be permanently stuck in the collision hash, causing a use-after-free once it is destroyed, because actors that have bCollideActors set to false will not be removed from the collision hash (because they are expected to not be present within it to begin with).